refactor: config 파싱 이후 WebServ 객체생성
