+++ src/HttpServer.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "Config.hpp"
 #include <cstddef>
 #include <functional>
 #include <string>
@@ -7,24 +8,6 @@
 #include <vector>
 
 using namespace std;
-
-struct RouteConfig {
-  bool isAutoIndex;
-  string httpRedirection;
-  string rootPath;
-  string index; // default file to answer if the request is a directory
-  string fileExtension;
-  string fastcgiPass;
-};
-
-struct Config {
-  string serverName;
-  string serverHost;
-  size_t serverPort;
-  string defaultErrorPage;
-  size_t maxClientBodySize;
-  unordered_map<string, RouteConfig> routeConfig;
-};
 
 class HttpServer {
 public:
